+++ backend/src/main/java/endolphin/backend/domain/auth/AuthController.java
@@ -1,7 +1,6 @@
 package endolphin.backend.domain.auth;
 
 import endolphin.backend.domain.auth.dto.OAuthResponse;
-import endolphin.backend.domain.auth.dto.UrlResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -16,13 +15,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
     private final AuthService authService;
-
-    @Operation(summary = "구글 로그인 URL", description = "구글 로그인 URL을 반환합니다.")
-    @GetMapping("/api/v1/google")
-    public ResponseEntity<UrlResponse> loginUrl() {
-        UrlResponse response = authService.getGoogleLoginUrl();
-        return ResponseEntity.ok(response);
-    }
 
     @Operation(summary = "구글 로그인 콜백", description = "사용자가 Google 계정으로 로그인하여 JWT 토큰을 발급받습니다.")
     @GetMapping("/oauth2/callback")
